requirements.txt
# ultralytics>=8.0.0
ultralytics>=8.0.200
# supervision for ByteTrack and annotations
supervision>=0.16.0
# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
# Computer vision
opencv-python>=4.8.0
numpy>=1.24.0
# Hugging Face datasets for Chicks4FreeID
datasets>=2.14.0
# Video export (optional)
moviepy>=1.0.3
# Utilities
tqdm>=4.65.0
Pillow>=10.0.0
